Add included crystals in the CALIFA geometry and s444 CALIFA geo
The geometry generation macro for CALIFA requires now as input a list of
the installed crystals and it will only create those alveoli that contain
the four crystals. A new file with the list of crystals installed in Nov 2019
is now included. A separate file includes the translations performed according
to the preliminary alignment of s444 experiment.